refactor: move user_repo_from_url utils -> builtin_plugins/github
