+++ src/builtin_plugins/github.rs
@@ -10,6 +10,7 @@ use hyper::net::HttpsConnector;
 use hyper::Client;
 use hyper_native_tls::NativeTlsClient;
 use serde::Deserialize;
+use url::{ParseError, Url};
 
 use crate::config::CfgMapExt;
 use crate::plugin::proto::{
@@ -119,7 +120,7 @@ impl PluginInterface for GithubPlugin {
         let remote = repo.find_remote(&cfg.remote)?;
         let remote_url = remote.url().ok_or(GithubPluginError::GitRemoteUndefined)?;
 
-        let (derived_name, derived_repo) = crate::utils::user_repo_from_url(remote_url)?;
+        let (derived_name, derived_repo) = user_repo_from_url(remote_url)?;
 
         let user = cfg.user.as_ref().unwrap_or(&derived_name);
         let repo_name = cfg.repository.as_ref().unwrap_or(&derived_repo);
@@ -256,6 +257,105 @@ impl Asset {
 
     pub fn content_type(&self) -> &str {
         &self.content_type
+    }
+}
+
+pub fn user_repo_from_url(url: &str) -> Result<(String, String), failure::Error> {
+    let path = match Url::parse(url) {
+        Err(ParseError::RelativeUrlWithoutBase) => match url.rfind(':') {
+            None => return Err(failure::err_msg("Can't parse path from remote URL")),
+            Some(colon_pos) => Some(
+                url[colon_pos + 1..]
+                    .split('/')
+                    .map(|s| s.to_owned())
+                    .collect::<Vec<_>>(),
+            ),
+        },
+        Err(_) => return Err(failure::err_msg("Can't parse remote URL")),
+        Ok(url) => url
+            .path_segments()
+            .map(|path| path.map(|seg| seg.to_owned()).collect::<Vec<_>>()),
+    };
+
+    let path = match path {
+        Some(ref path) if path.len() == 2 => path,
+        _ => {
+            return Err(failure::err_msg(
+                "Remote URL should contain user and repository",
+            ))
+        }
+    };
+
+    let user = path[0].clone();
+    let repo = match path[1].rfind(".git") {
+        None => path[1].clone(),
+        Some(suffix_pos) => {
+            let valid_pos = path[1].len() - 4;
+            if valid_pos == suffix_pos {
+                let path = &path[1][0..suffix_pos];
+                path.into()
+            } else {
+                path[1].clone()
+            }
+        }
+    };
+
+    Ok((user, repo))
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn parses_remote_urls() {
+        let urls = [
+            "https://github.com/user/repo.git",
+            "https://github.com/user/repo",
+            "git@github.com:user/repo.git",
+            "git@github.com:user/repo",
+            "ssh://github.com/user/repo",
+            "ssh://github.com/user/repo.git",
+        ];
+
+        for url in &urls {
+            println!("Testing '{:?}'", url);
+            let (user, repo) = user_repo_from_url(url).unwrap();
+
+            assert_eq!("user", user);
+            assert_eq!("repo", repo);
+        }
+    }
+
+    #[test]
+    fn parses_other_urls() {
+        let urls = [(
+            "https://github.com/user/repo.git.repo",
+            "user",
+            "repo.git.repo",
+        )];
+
+        for &(url, exp_user, exp_repo) in &urls {
+            println!("Testing '{:?}'", url);
+            let (user, repo) = user_repo_from_url(url).unwrap();
+
+            assert_eq!(exp_user, user);
+            assert_eq!(exp_repo, repo);
+        }
+    }
+
+    #[test]
+    fn fail_some_urls() {
+        let urls = [
+            "https://github.com/user",
+            "https://github.com/user/repo/issues",
+            "://github.com/user/",
+        ];
+
+        for url in &urls {
+            println!("Testing '{:?}'", url);
+            assert!(user_repo_from_url(url).is_err());
+        }
     }
 }
